[Mailer] Fix BC breaking event name change
