+++ src/Symfony/Component/Mailer/Event/MessageEvent.php
@@ -12,6 +12,8 @@
 namespace Symfony\Component\Mailer\Event;
 
 use Symfony\Component\Mailer\Envelope;
+use Symfony\Component\Mailer\Exception\LogicException;
+use Symfony\Component\Messenger\Stamp\StampInterface;
 use Symfony\Component\Mime\RawMessage;
 use Symfony\Contracts\EventDispatcher\Event;
 
@@ -20,12 +22,15 @@ use Symfony\Contracts\EventDispatcher\Event;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class MessageEvent extends Event
+final class MessageEvent extends Event
 {
     private RawMessage $message;
     private Envelope $envelope;
     private string $transport;
     private bool $queued;
+
+    /** @var StampInterface[] */
+    private array $stamps = [];
 
     public function __construct(RawMessage $message, Envelope $envelope, string $transport, bool $queued = false)
     {
@@ -63,5 +68,26 @@ class MessageEvent extends Event
     public function isQueued(): bool
     {
         return $this->queued;
+    }
+
+    public function addStamp(StampInterface $stamp): void
+    {
+        if (!$this->queued) {
+            throw new LogicException(sprintf('Cannot call "%s()" on a message that is not meant to be queued', __METHOD__));
+        }
+
+        $this->stamps[] = $stamp;
+    }
+
+    /**
+     * @return StampInterface[]
+     */
+    public function getStamps(): array
+    {
+        if (!$this->queued) {
+            throw new LogicException(sprintf('Cannot call "%s()" on a message that is not meant to be queued.', __METHOD__));
+        }
+
+        return $this->stamps;
     }
 }
