+++ src/Symfony/Component/Mailer/EventListener/MessageLoggerListener.php
@@ -14,7 +14,6 @@ namespace Symfony\Component\Mailer\EventListener;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Mailer\Event\MessageEvent;
 use Symfony\Component\Mailer\Event\MessageEvents;
-use Symfony\Component\Mailer\Event\QueuingMessageEvent;
 use Symfony\Contracts\Service\ResetInterface;
 
 /**
@@ -50,7 +49,6 @@ class MessageLoggerListener implements EventSubscriberInterface, ResetInterface
     {
         return [
             MessageEvent::class => ['onMessage', -255],
-            QueuingMessageEvent::class => ['onMessage', -255],
         ];
     }
 }
