+++ src/Symfony/Component/Mailer/Tests/EventListener/MessengerTransportListenerTest.php
@@ -13,7 +13,7 @@ namespace Symfony\Component\Mailer\Tests\EventListener;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Mailer\Envelope;
-use Symfony\Component\Mailer\Event\QueuingMessageEvent;
+use Symfony\Component\Mailer\Event\MessageEvent;
 use Symfony\Component\Mailer\EventListener\MessengerTransportListener;
 use Symfony\Component\Messenger\Stamp\TransportNamesStamp;
 use Symfony\Component\Mime\Address;
@@ -27,7 +27,7 @@ class MessengerTransportListenerTest extends TestCase
         $l = new MessengerTransportListener();
         $envelope = new Envelope(new Address('sender@example.com'), [new Address('recipient@example.com')]);
         $message = new Message(new Headers());
-        $event = new QueuingMessageEvent($message, $envelope, 'smtp');
+        $event = new MessageEvent($message, $envelope, 'smtp', true);
         $l->onMessage($event);
         $this->assertEmpty($event->getStamps());
     }
@@ -38,7 +38,7 @@ class MessengerTransportListenerTest extends TestCase
         $envelope = new Envelope(new Address('sender@example.com'), [new Address('recipient@example.com')]);
         $headers = (new Headers())->addTextHeader('X-Bus-Transport', 'async');
         $message = new Message($headers);
-        $event = new QueuingMessageEvent($message, $envelope, 'smtp');
+        $event = new MessageEvent($message, $envelope, 'smtp', true);
         $l->onMessage($event);
         $this->assertCount(1, $event->getStamps());
         /* @var TransportNamesStamp $stamp */
@@ -54,7 +54,7 @@ class MessengerTransportListenerTest extends TestCase
         $name = 'söme_very_long_and_weïrd transport name-for-messenger!';
         $headers = (new Headers())->addTextHeader('X-Bus-Transport', ' async , async1,'.$name);
         $message = new Message($headers);
-        $event = new QueuingMessageEvent($message, $envelope, 'smtp');
+        $event = new MessageEvent($message, $envelope, 'smtp', true);
         $l->onMessage($event);
         $this->assertCount(1, $event->getStamps());
         /* @var TransportNamesStamp $stamp */
